requirements.txt
python-telegram-bot==13.15
urllib3==1.26.15
environs==14.1.1
standard-imghdr==3.13.0
redis~=6.1.0